refactor(config): Extract shared Tailwind config to packages/config
Consolidate ~850 lines of duplicated Tailwind configuration from
apps/blog and apps/storybook into a single shared preset.

The shared config provides:
- Semantic color system (ground, figure, border, action, status, etc.)
- Typography tokens with font-size/weight/line-height/letter-spacing
- Spacing system (stack, inline, inset, gutter, section)
- Motion tokens (durations, easing)
- 4 custom plugins: typography, spacing, motion, accessibility

Apps now use the preset pattern:
  const sharedConfig = require('@blog/config/tailwind.config');
  module.exports = { presets: [sharedConfig], content: [...] };

- Blog tests: 53/53 passing
- Blog build: successful
- Storybook build: successful
